+++ apps/io-func-wallet-solution-backend/src/app/config.ts
@@ -3,7 +3,7 @@ import * as t from "io-ts";
 import { pipe } from "fp-ts/lib/function";
 import * as RE from "fp-ts/lib/ReaderEither";
 import { getFederationEntityConfigFromEnvironment } from "../infra/federation-entity/config";
-import { FederationEntityMetadata } from "../wallet-provider";
+import { FederationEntityMetadata } from "../entity-configuration";
 import {
   CryptoConfiguration,
   getCryptoConfigFromEnvironment,
